fix!: raise error if log_location config. parameter is empty
This introduce new behavior when log_location config. parameter is empty. Previously temporary directory is used when empty log location is specified, but it causes data loss when the temporary dir. is not saved and the behavior looks confusing to users. So starting Tsurugi DB will fail with empty log location.

https://github.com/project-tsurugi/tsurugi-issues/issues/549