fix(limestone): handle 0-filled WAL tail corruption via format v6 upgrade
This commit addresses issue #1257:
https://github.com/project-tsurugi/tsurugi-issues/issues/1257

Unexpected 0-filled data may be appended to the tail of a WAL file due
to power loss or abrupt termination, which previously caused the file
to be treated as unrecoverably corrupted.

To address this, the WAL file format is upgraded from version 5 to 6.

Version 6 introduces a `marker_end` entry at the end of each epoch,
enabling the system to distinguish valid data from trailing garbage.

Key changes:
- Files with format version <6 are migrated to version 6 at startup.
- WAL rotation is performed during migration to isolate old-format files.
- Version 6 WALs are not readable by systems using version 5 or earlier.

As a result, WALs with trailing 0-filled data are now considered
recoverable and can be repaired using `tglogutil`.